Try out adding/removing tasks using Vuex
Had to make a minor change to a mutation: don't wrap the `task` parameter in braces.

This is the example given by the Nuxt instructions, but all it was doing was causing the `remove` method to remove the last entry in the list instead of the one corresponding with the delete button.